requirements.txt
Django>=4.2
djangorestframework
psycopg2
Pillow
djangorestframework-simplejwt